Add Functor and Applicative instances to all Monads
These are the minimum required fixes to get greencard building under GHC
7.10.  As noted in the comments around the existing Functor instances,
though, this may cause compatibility issues with earlier compilers and
indeed other, non-GHC Haskell 98 compilers.

I am submitting this commit not with a view to having it accepted as is
but as a starting-point for a conversation about how to move forward
with supporting the latest GHC alongside earlier versions.